backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
# pinned to versions known to be compatible in this environment
openai==1.3.0
httpx==0.23.3

